chore (refs T22614): move DpTooltipIcon to client/js.
