+++ demosplan/DemosPlanProcedureBundle/Resources/client/js/bundles/administrationMemberList.js
@@ -11,7 +11,7 @@
  * This is the entrypoint for administration_member_list.html.twig
  */
 import { addFormHiddenField, removeFormHiddenField } from '@DpJs/lib/FormActions'
-import DpTooltipIcon from '@DemosPlanCoreBundle/components/DpTooltipIcon'
+import DpTooltipIcon from '@DpJs/components/core/DpTooltipIcon'
 import { initialize } from '@DemosPlanCoreBundle/InitVue'
 
 const components = {
