+++ demosplan/DemosPlanStatementBundle/Resources/client/js/bundles/listTags.js
@@ -12,7 +12,7 @@
  */
 
 import AnimateById from '@DpJs/lib/AnimateById'
-import DpTooltipIcon from '@DemosPlanCoreBundle/components/DpTooltipIcon'
+import DpTooltipIcon from '@DpJs/components/core/DpTooltipIcon'
 import DpUploadFiles from '@DpJs/components/core/DpUpload/DpUploadFiles'
 import { initialize } from '@DemosPlanCoreBundle/InitVue'
